This PR Improves and simplifies the non dimensionalization of `AbstractMaterialParam`s, so that it does not rely on `Setfields.jl` anymore. This solves the problems reported by @Iddingsite in #241. It also slighlty improves the performance

```julia
using GeoParams
# This tests the MaterialParameters structure
CharUnits_GEO = GEO_units(; viscosity=1Pa * s, length=1m)
x = DiffusionCreep()
```

This PR:
```julia-repl
julia> @benchmark nondimensionalize($x, $CharUnits_GEO)
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.470 μs …  18.470 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.620 μs               ┊ GC (median):    0.00%    
 Time  (mean ± σ):   1.621 μs ± 307.944 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

       ▁ ▅█▇ ▃           ▁ ▄▄ ▁
  ▂▃▄▁██▁███▁█▇▁▄▃▃▁▃▄▁▅▆█▁██▁█▆▅▁▅▄▁▅▅▅▁▅▄▁▃▃▃▁▃▂▁▂▂▂▁▂▂▁▂▂▂ ▃
  1.47 μs         Histogram: frequency by time        1.89 μs <        

 Memory estimate: 688 bytes, allocs estimate: 33.
```

`#main`:
```julia-repl
julia> @benchmark nondimensionalize($x, $CharUnits_GEO)
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.740 μs …   5.230 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.810 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.831 μs ± 142.881 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

    █▇▄
  ▂▄███▆▆▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▁▂▁▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▂▁▁▂▂▂▁▁▂▂▂▂ ▃
  1.74 μs         Histogram: frequency by time        2.81 μs <

 Memory estimate: 688 bytes, allocs estimate: 33.
 ```